+++ app/client/src/routes/home.tsx
@@ -18,7 +18,7 @@ import CopyBox from 'components/copyBox';
 import GlossaryPanel, { GlossaryTerm } from 'components/glossaryPanel';
 import InfoTooltip from 'components/infoTooltip';
 import { Loading } from 'components/loading';
-import Modal from 'components/modal';
+import DownloadModal from 'components/downloadModal';
 import RadioButtons from 'components/radioButtons';
 import SourceSelect from 'components/sourceSelect';
 import Summary from 'components/summary';
@@ -29,7 +29,6 @@ import {
   fields as allFields,
   getData,
   options as listOptions,
-  postData,
   profiles,
   serverUrl,
 } from 'config';
@@ -83,16 +82,6 @@ type UrlQueryParam = [string, Primitive];
 
 type UrlQueryState = {
   [field: string]: Primitive | Primitive[];
-};
-
-type PostData = {
-  filters: {
-    [field: string]: Primitive | Primitive[];
-  };
-  options: {
-    f?: string;
-    format?: string;
-  };
 };
 
 /*
@@ -1005,112 +994,5 @@ function FilterFields({ handlers, state, staticOptions }: FilterFieldsProps) {
         </div>
       ))}
     </div>
-  );
-}
-
-type DownloadModalProps = {
-  downloadStatus: Status;
-  filename: string | null;
-  onClose: () => void;
-  queryData: PostData;
-  queryUrl: string | null;
-  setDownloadStatus: (status: Status) => void;
-};
-
-function DownloadModal({
-  downloadStatus,
-  filename,
-  onClose,
-  queryData,
-  queryUrl,
-  setDownloadStatus,
-}: DownloadModalProps) {
-  const [count, setCount] = useState<number | null>(null);
-  const [countStatus, setCountStatus] = useState<Status>('idle');
-
-  // Get the row count for the current query
-  useEffect(() => {
-    if (!queryUrl) return;
-
-    const countUrl = `${queryUrl}/count`;
-    setCountStatus('pending');
-    postData(countUrl, queryData)
-      .then((res) => {
-        setCount(parseInt(res.count));
-        setCountStatus('success');
-      })
-      .catch((err) => {
-        console.error(err);
-        setCountStatus('failure');
-      });
-  }, [queryData, queryUrl]);
-
-  // Retrieve the requested data in the specified format
-  const executeQuery = useCallback(() => {
-    if (!queryUrl) return;
-    if (!filename) return;
-
-    setDownloadStatus('pending');
-    postData(queryUrl, queryData, 'blob')
-      .then((res) => {
-        const fileUrl = window.URL.createObjectURL(res);
-        const trigger = document.createElement('a');
-        trigger.style.display = 'none';
-        trigger.href = fileUrl;
-        trigger.download = filename;
-        trigger.click();
-        window.URL.revokeObjectURL(fileUrl);
-        setDownloadStatus('success');
-      })
-      .catch((err) => {
-        console.error(err);
-        setDownloadStatus('failure');
-      })
-      .finally(() => onClose());
-  }, [queryUrl, filename, setDownloadStatus, queryData, onClose]);
-
-  return (
-    <Modal onClose={onClose}>
-      <h2 className="usa-modal__heading">Download Status</h2>
-      {countStatus === 'pending' && (
-        <div className="usa-prose">
-          <p>Validating query, please wait...</p>
-        </div>
-      )}
-      {countStatus === 'failure' && (
-        <Alert type="error">
-          The specified query could not be executed at this time.
-        </Alert>
-      )}
-      {countStatus === 'success' && (
-        <>
-          <div className="usa-prose">
-            <p>
-              Your query will return <strong>{count}</strong> rows.
-            </p>
-            <p>Click continue to download the data.</p>
-          </div>
-          <div className="usa-modal__footer">
-            <ul className="flex-justify-center usa-button-group">
-              <li className="usa-button-group__item">
-                <button type="button" className="usa-button" onClick={onClose}>
-                  Cancel
-                </button>
-              </li>
-              <li className="usa-button-group__item">
-                <button
-                  className="usa-button"
-                  disabled={count === 0}
-                  onClick={executeQuery}
-                  type="button"
-                >
-                  {downloadStatus === 'pending' ? 'Working...' : 'Continue'}
-                </button>
-              </li>
-            </ul>
-          </div>
-        </>
-      )}
-    </Modal>
   );
 }
